pyproject.toml
[tool.poetry]
name = "sdec"
version = "0.1.0"
description = "Sun Devil Rocketry embedded controller CLI."
authors = ["Caden Kriese <13630061+cadenkriese@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.1"
rich = "^10.2.2"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pylint = "^2.10.2"
black = "^21.9b0"
pytest-xdist = "^2.4.0"

[tool.poetry.scripts]
sdec = "sdec.sdec:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
